pyproject.toml
[project]
name = "jarvis"
version = "0.1.0"
description = "FRIDAY — Personal AI Operating System"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.2",
    "chromadb>=1.5.5",
    "google-api-python-client>=2.193.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    "httpx>=0.28.1",
    "ollama>=0.6.1",
    "playwright>=1.58.0",
    "prompt-toolkit>=3.0.52",
    "python-dotenv>=1.2.2",
    "pywebostv>=0.8.9",
    "rich>=14.3.3",
    "tavily-python>=0.7.23",
    "wakeonlan>=3.1.0",
    "weasyprint>=63.0",
    "jinja2>=3.1.6",
    "pypdf>=6.9.1",
    "pdfplumber>=0.11.9",
    "tweepy>=4.16.0",
    "sounddevice>=0.5.1",
    "numpy>=1.26.0",
    "mlx-whisper>=0.4.0",
    "kokoro-onnx>=0.4.0",
    "onnxruntime>=1.19.0",
]

[project.scripts]
friday = "friday.cli:run"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["friday*"]
